requirements.txt
setuptools~=72.1.0
PyYAML~=5.4.1
types-PyYAML~=6.0.5
openpyxl
openpyxl-stubs